ask_your_document/requirements.txt
streamlit==1.28.0
chromadb==0.4.15
openai==1.3.0
python-dotenv==1.0.0
tiktoken==0.5.1

